+++ src/main/java/frc/robot/commands/arm/ArcadeArm.java
@@ -2,50 +2,62 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
-
-import edu.wpi.first.math.controller.PIDController;
+import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.constants.ArmConstants;
 import frc.robot.subsystems.Arm;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ArmToPos extends Command {
-  private double m_pos;
-  private Arm m_arm;
-  private double m_speed;
+public class ArcadeArm extends Command {
   
-  // Creates a PIDController with gains kP, kI, and kD
-  PIDController m_pid = new PIDController(ArmConstants.kP, ArmConstants.kI, ArmConstants.kD);
-  
-  /** Creates a new ArmToPos. */
-  // NOTE
-  // max encoder ticks the arm should be at (near scoring position) == 0.904
-  // min encoder ticks the arm should be at (at the bottom, when intaking fron the indexer) == 0.469
-  public ArmToPos(Arm arm, double pos) {
-    m_pos = pos;
+  /** Creates a new ArcadeArm. */
+
+private Arm m_arm;
+private Joystick m_operatorJoystick;
+private double m_speed;
+
+  public ArcadeArm(Arm arm, Joystick operatorjoystick) {
     m_arm = arm;
-    addRequirements(m_arm); 
+    m_operatorJoystick = operatorjoystick;
+  
     // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(m_arm);
   }
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_speed = -m_pid.calculate(m_arm.getEncoderPosition(), m_pos);
+    m_speed = m_operatorJoystick.getRawAxis(ArmConstants.kArmManualControlAxis) * ArmConstants.kArmMultiplier;
     m_arm.setSpeed(m_speed);
-    SmartDashboard.putNumber("Real Arm Position", m_arm.getEncoderPosition());
-    SmartDashboard.putNumber("Arm Set Position", m_pos);
-    SmartDashboard.putNumber("Arm Error", m_pos - m_arm.getEncoderPosition());
-    SmartDashboard.putNumber("Arm PID Speed", -m_speed);
-  }
-
+    /*
+    if (speed>0){
+      if (m_arm.getTopLimitSwitch()){
+        SmartDashboard.putNumber("Arm Speed", 0);
+        m_arm.setSpeed(0);
+      } else {
+        SmartDashboard.putNumber("Arm Speed", speed);
+        m_arm.setSpeed(speed);
+      }
+    } else {
+      if (m_arm.getBottomLimitSwitch()){
+        SmartDashboard.putNumber("Arm Speed", 0);
+        m_arm.setSpeed(0);
+      } else {
+        SmartDashboard.putNumber("Arm Speed", speed);
+        m_arm.setSpeed(speed);
+      }
+    }
+      */
+    }
+    
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
@@ -56,7 +68,5 @@ public class ArmToPos extends Command {
   @Override
   public boolean isFinished() {
     return false;
-    // TODO
-    // NEED THE FINISH PID THING
   }
 }
