+++ src/main/java/frc/robot/commands/elevator/ElevatorRatioTest.java
@@ -2,44 +2,37 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.elevator;
 
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants.IntakeConstants;
-import frc.robot.subsystems.Indexer;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.Elevator;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class RunIndexer extends Command {
- //change later
-  private Indexer m_indexer;
-  private double m_speed; 
-
-  /** Creates a new SpinForIntake. */
-  public RunIndexer(Indexer indexer, double speed) {
-    m_indexer = indexer;
+public class ElevatorRatioTest extends Command {
+  Elevator m_elevator;
+  double m_speed;
+  /** Creates a new ElevatorRatioTest. */
+  public ElevatorRatioTest(Elevator elevator, double speed) {
+    m_elevator = elevator;
     m_speed = speed;
-    // Use addRequirements() here to declare subsystem dependencies.
-    addRequirements(m_indexer);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    m_indexer.setIndexerSpeed(0);
+
   }
+
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    m_indexer.setIndexerSpeed(m_speed);
-    SmartDashboard.putNumber("Indexer Speed", m_speed);
+    m_elevator.setSpeed(m_speed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    m_indexer.setIndexerSpeed(0);
+    m_elevator.setSpeed(0);
   }
 
   // Returns true when the command should end.
@@ -48,4 +41,3 @@ public class RunIndexer extends Command {
     return false;
   }
 }
-
