+++ src/main/java/frc/robot/commands/intake_indexer/RunIntake.java
@@ -2,28 +2,44 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.intake_indexer;
 
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.Constants.IntakeConstants;
+import frc.robot.subsystems.Intake;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ArmPID extends Command {
-  /** Creates a new ArmPID. */
-  public ArmPID() {
+public class RunIntake extends Command {
+ //change later
+  private Intake m_intakeMotor;
+  private double m_speed; 
+
+  /** Creates a new SpinForIntake. */
+  public RunIntake(Intake intake, double speed) {
+    m_intakeMotor = intake;
+    m_speed = speed;
     // Use addRequirements() here to declare subsystem dependencies.
+    addRequirements(m_intakeMotor);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
-
+  public void initialize() {
+    m_intakeMotor.setIntakeSpeed(0);
+  }
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    m_intakeMotor.setIntakeSpeed(m_speed);
+    SmartDashboard.putNumber("Intake Speed", m_speed);
+  }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    m_intakeMotor.setIntakeSpeed(0);
+  }
 
   // Returns true when the command should end.
   @Override
@@ -31,3 +47,4 @@ public class ArmPID extends Command {
     return false;
   }
 }
+
