+++ src/main/java/frc/robot/commands/intake_indexer/RunIntakeWithIndexerJoystick.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.intake_indexer;
 
 import java.nio.file.WatchEvent.Kind;
