+++ src/main/java/frc/robot/commands/pivot/PivotToPosition.java
@@ -2,24 +2,24 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.pivot;
 
-import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.PivotConstants;
 import frc.robot.subsystems.Pivot;
 
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class ArcadePivot extends Command {
+public class PivotToPosition extends Command {
   private Pivot m_pivot;
-  private Joystick m_joystick;
-  private double speed;
+  private double m_speed;
+  private double m_goal;
 
   /** Creates a new ArcadePivot. */
-  public ArcadePivot(Pivot pivot, Joystick joystick) {
+  public PivotToPosition(Pivot pivot, double speed, double goal) {
     m_pivot = pivot;
-    m_joystick = joystick;
+    m_speed = speed;
+    m_goal = goal;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(m_pivot);
   }
@@ -33,25 +33,12 @@ public class ArcadePivot extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    speed = m_joystick.getRawAxis(PivotConstants.kManualPivotControlAxis);
+    m_pivot.setPivotSpeed(m_speed);
+    SmartDashboard.putNumber("Pivot Speed", m_speed);
     SmartDashboard.putNumber("Rotations", m_pivot.getPivotPosition());
-    SmartDashboard.putBoolean("Bottom Pivot Limit", m_pivot.getBottomLimitSwitch());
-    SmartDashboard.putBoolean("Top Pivot Limit", m_pivot.getTopLimitSwitch());
     
-    // if our pivot is at the top
-    if ((speed < 0) && ((m_pivot.getTopLimitSwitch() || (m_pivot.getPivotPosition() < PivotConstants.kTopPivotRotations)))){
-      m_pivot.setPivotSpeed(0);
-      SmartDashboard.putNumber("Pivot Speed", 0);
-
-      // if our pivot is at the bottom
-    } else if ((speed > 0) && ((m_pivot.getBottomLimitSwitch() || (m_pivot.getPivotPosition() > PivotConstants.kBottomPivotRotations)))){
-      m_pivot.setPivotSpeed(0);
-      SmartDashboard.putNumber("Pivot Speed", 0);
-
-    } else {
-      m_pivot.setPivotSpeed(speed);
-      SmartDashboard.putNumber("Pivot Speed", speed);
-    }
+    //m_pivot.getPivotSim().setInputVoltage(12 * m_speed);
+    //m_pivot.getPivotSim().update(0.02);
   }
 
   // Called once the command ends or is interrupted.
@@ -63,6 +50,9 @@ public class ArcadePivot extends Command {
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
+    if((m_pivot.getPivotPosition() - m_goal > -PivotConstants.kRotationTolerance) && (m_pivot.getPivotPosition() - m_goal < PivotConstants.kRotationTolerance)) {
+      return true;
+    }
     return false;
   }
 }
